+++ crates/tests/riddle/tests/struct.idl
@@ -1,5 +1,5 @@
 mod TypeNamespace {
-    // struct TypeName {
-    //     F1: u32,
-    // }
+    struct TypeName {
+        F1: u32,
+    }
 }
