+++ crates/tests/riddle/tests/struct.rs
@@ -7,4 +7,6 @@ fn riddle_struct() {
     let files = File::with_default(&[&output]).expect("Failed to open winmd files");
     let reader = &Reader::new(&files);
 
+    let def = reader.get(TypeName::new("TypeNamespace", "TypeName")).next().expect("Type missing");
+    assert_eq!(reader.type_def_kind(def), TypeKind::Struct);
 }
